Fixing various issues to get the key/value graph store to work
