+++ gdbi/tpath/fields.go
@@ -1,0 +1,5 @@
+package tpath
+
+func IsGraphField(f string) bool {
+	return f == "_gid" || f == "_label" || f == "_to" || f == "_from"
+}
